Remove BOARD_SYSTEMSDK_VERSIONS for SUN
The value automatically set by the build system is more correct. Setting
it manually to 33 actually was erronous because there are some modules
installed to product or vendor partition which have `sdk_version` set to
`system_current` which is VIC as of now. Those errors were not detected
due to a bug in the build system, but that is now detected after the
refactoring done as part of https://android-review.git.corp.google.com/q/topic:%22limit_systemsdk%22